优化

+++ app/src/main/java/com/web3/airdrop/data/TaskConfig.kt
@@ -5,6 +5,6 @@ import com.web3.airdrop.base.IPanelTaskModule
 data class TaskConfig(
     val globalMode: Boolean = false,//全局模式
     val timingMode: Boolean = false,//定时模式
-    val timing: Int = 0,//每隔小时执行
+    val timing: Long = 0,//每隔小时执行
     val taskList: List<IPanelTaskModule.PanelTask>? //任务
 )
